Move testing application into .NET Framework project
Some dependancies are not supported by Core, so the testing app must be in Framework.
