+++ Developer Samples/CommandLineTools/CommandLineTools.sln
@@ -7,7 +7,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PromoteRuleApp", "PromoteRu
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BuildIrJsRuleApp", "BuildIrJsRuleApp\BuildIrJsRuleApp.csproj", "{23CF13F1-F24F-4C0A-8299-E95683A53435}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RunTests", "RunTests\RunTests.csproj", "{E7CB8304-B892-4F7C-9B8F-71602E04B238}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExecuteTests", "ExecuteTests\ExecuteTests.csproj", "{9ABF0D22-46AD-49DE-B122-3180FDA2F391}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -23,10 +23,10 @@ Global
 		{23CF13F1-F24F-4C0A-8299-E95683A53435}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{23CF13F1-F24F-4C0A-8299-E95683A53435}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{23CF13F1-F24F-4C0A-8299-E95683A53435}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E7CB8304-B892-4F7C-9B8F-71602E04B238}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E7CB8304-B892-4F7C-9B8F-71602E04B238}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E7CB8304-B892-4F7C-9B8F-71602E04B238}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E7CB8304-B892-4F7C-9B8F-71602E04B238}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9ABF0D22-46AD-49DE-B122-3180FDA2F391}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9ABF0D22-46AD-49DE-B122-3180FDA2F391}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9ABF0D22-46AD-49DE-B122-3180FDA2F391}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9ABF0D22-46AD-49DE-B122-3180FDA2F391}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
