+++ Developer Samples/CommandLineTools/ExecuteTests/packages.config
@@ -1,0 +1,11 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="InRule.Authoring.Core" version="5.4.0" targetFramework="net461" />
+  <package id="InRule.Authoring.SDK" version="5.4.0" targetFramework="net461" />
+  <package id="InRule.Common" version="5.4.0" targetFramework="net461" />
+  <package id="InRule.Repository" version="5.4.0" targetFramework="net461" />
+  <package id="InRule.Runtime" version="5.4.0" targetFramework="net461" />
+  <package id="Microsoft.AspNet.WebApi.Client" version="5.2.3" targetFramework="net461" />
+  <package id="Mono.Options" version="5.3.0.1" targetFramework="net461" />
+  <package id="Newtonsoft.Json" version="12.0.1" targetFramework="net461" />
+</packages>
