--- Developer Samples/CommandLineTools/RunTests/RunTests.csproj
@@ -1,33 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Mono.Options" Version="5.3.0.1" />
-    <PackageReference Include="System.Configuration.ConfigurationManager" Version="4.5.0" />
-    <PackageReference Include="System.IO.Packaging" Version="4.5.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Reference Include="InRule.Common">
-      <HintPath>..\..\..\..\..\..\..\..\..\Program Files (x86)\InRule\irSDK\bin\InRule.Common.dll</HintPath>
-    </Reference>
-    <Reference Include="InRule.Repository">
-      <HintPath>..\..\..\..\..\..\..\..\..\Program Files (x86)\InRule\irSDK\bin\InRule.Repository.dll</HintPath>
-    </Reference>
-    <Reference Include="InRule.Runtime">
-      <HintPath>..\..\..\..\..\..\..\..\..\Program Files (x86)\InRule\irSDK\bin\InRule.Runtime.dll</HintPath>
-    </Reference>
-    <Reference Include="InRule.Runtime.Testing">
-      <HintPath>..\..\..\..\..\..\..\..\..\Program Files (x86)\InRule\irSDK\bin\InRule.Runtime.Testing.dll</HintPath>
-    </Reference>
-  </ItemGroup>
-
-  <Target Name="PostBuild" AfterTargets="PostBuildEvent">
-    <Exec Command="rem dotnet publish -c Debug -r win10-x64" />
-  </Target>
-
-</Project>
